Using implementation and exception creation
