+++ src/main/java/com/eddu/java_tcs1/exceptions/ResourceNotFoundException.java
@@ -1,0 +1,13 @@
+package com.eddu.java_tcs1.exceptions;
+
+public class ResourceNotFoundException extends Exception {
+
+	
+	private static final long serialVersionUID = 1L;
+		
+	public ResourceNotFoundException (String message) {
+		
+		super(message);
+	}
+	
+}
