+++ src/main/java/com/eddu/java_tcs1/services/AccountService.java
@@ -2,35 +2,15 @@ package com.eddu.java_tcs1.services;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
+import com.eddu.java_tcs1.exceptions.ResourceNotFoundException;
 import com.eddu.java_tcs1.model.Account;
-import com.eddu.java_tcs1.repository.AccountRepository;
 
-@Service
-public class AccountService {
-	@Autowired
-	private AccountRepository accountRepository;
+public interface AccountService {
 
-	public List<Account> getAllAccounts() {
-		return accountRepository.findAll();
-	}
-
-	public Account getAccountId(Integer Id) {
-		return accountRepository.findById(Id).get();
-	}
-
-	public Account saveOrUpdateAccount(Account account) {
-		return  accountRepository.save(account);
-	}
-
-	public void deleteAccount(Integer id) {
-		accountRepository.deleteById(id);
-	}
-
-	public void updateAccount(Account account, Integer id) {
-		accountRepository.save(account);
-	}
-
+	public List<Account> getAllAccounts();
+	public Account getAccountId(Integer Id);
+	public Account saveOrUpdateAccount(Account account);
+	public void deleteAccount(Integer id) throws ResourceNotFoundException;
+	public void updateAccount(Account account, Integer id) throws ResourceNotFoundException;
+	
 }
